Add files via upload
updates for 2.3 are all about making first use easier.  we have done away with the need to a tasks.csv as this was just a list of numbers.  we now have a real list of rounds of the championship in the championship file and on first run the code creates the dir that will store the igc and other files.